Issue new device IDs to replace invalid IDs.
Previously, `WorkerWS.handshake()` returned `ErrInvalidID` and closed
the connection if the client sent an invalid ID. The Gecko client does
not currently detect non-200 responses or clear invalid IDs before
reconnecting. This can trigger a reconnect loop if the
`services.push.userAgentID` preference is set to an invalid value by
a user (via `about:config`) or third-party Simple Push server (the
spec recommends, but does not require, UUIDs).

This commit relaxes the server behavior to issue a new, valid device
ID if the client supplies an invalid one.